Resolve import errors due to renaming of packing person to guest
